Functional, but just work in progress to spritification
mines and nobles are now sprites
Realized I could do a front and back of the mines in one class
But I'm concerned about adding all the mine sprites to the
gamesession groups and rendering a bunch of card backs on top of
each other. seems like a waste, so might create a 'deep stack'
group for hidden cards.